requirements.txt
pandas
numpy
scikit-learn
tqdm
python-dotenv
google-genai
fastapi
uvicorn
scipy
apscheduler
pymongo 
dnspython